Move to examples/llms
