--- docs/extras/modules/model_io/models/chat/integrations/bittensor.mdx
@@ -1,10 +1,0 @@
-# Bittensor
-
-LangChain.js supports Neural Internet's Bittensor LLM models. 
-
-Here's an example:
-
-import CodeBlock from "@theme/CodeBlock";
-import BittensorExample from "@examples/models/chat/bittensor.ts";
-
-<CodeBlock language="typescript">{BittensorExample}</CodeBlock>
