+++ docs/extras/modules/model_io/models/llms/integrations/ni_bittensor.mdx
@@ -1,0 +1,10 @@
+# NIBittensor
+
+LangChain.js offers experimental support for Neural Internet's Bittensor LLM models.
+
+Here's an example:
+
+import CodeBlock from "@theme/CodeBlock";
+import NIBittensorExample from "@examples/models/llms/ni_bittensor.ts";
+
+<CodeBlock language="typescript">{NIBittensorExample}</CodeBlock>
