+++ langchain/package.json
@@ -583,9 +583,9 @@
     "experimental/chat_models/anthropic_functions.cjs",
     "experimental/chat_models/anthropic_functions.js",
     "experimental/chat_models/anthropic_functions.d.ts",
-    "experimental/chat_models/bittensor.cjs",
-    "experimental/chat_models/bittensor.js",
-    "experimental/chat_models/bittensor.d.ts",
+    "experimental/llms/bittensor.cjs",
+    "experimental/llms/bittensor.js",
+    "experimental/llms/bittensor.d.ts",
     "evaluation.cjs",
     "evaluation.js",
     "evaluation.d.ts",
@@ -2053,10 +2053,10 @@
       "import": "./experimental/chat_models/anthropic_functions.js",
       "require": "./experimental/chat_models/anthropic_functions.cjs"
     },
-    "./experimental/chat_models/bittensor": {
-      "types": "./experimental/chat_models/bittensor.d.ts",
-      "import": "./experimental/chat_models/bittensor.js",
-      "require": "./experimental/chat_models/bittensor.cjs"
+    "./experimental/llms/bittensor": {
+      "types": "./experimental/llms/bittensor.d.ts",
+      "import": "./experimental/llms/bittensor.js",
+      "require": "./experimental/llms/bittensor.cjs"
     },
     "./evaluation": {
       "types": "./evaluation.d.ts",
